requirements.txt
Django==5.1.*
celery==5.4.*
djangorestframework==3.15.*